requirements.txt
argon2-cffi==21.1.0
asgiref==3.4.1
astroid==2.7.3
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bandit==1.7.0
black==21.12b0
bleach==4.1.0
boto3==1.20.34
botocore==1.23.34
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==4.0
django-allauth==0.47.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-mathfilters==1.0.0
django-phonenumber-field==6.0.0
django-storages==1.12.3
docutils==0.17.1
entrypoints==0.3
filelock==3.0.12
flake8==3.9.2
gitdb==4.0.7
GitPython==3.1.18
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
isort==5.9.3
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keyring==23.2.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy==0.910
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
oauthlib==3.1.1
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pbr==5.6.0
pep8==1.7.1
pexpect==4.8.0
phonenumbers==8.12.41
pickleshare==0.7.5
Pillow==9.0.0
pipenv==2021.5.29
pkginfo==1.7.1
platformdirs==2.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2==2.9.2
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.10.0
PyJWT==2.3.0
pylama==7.7.1
pylint==2.10.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python3-openid==3.2.0
PyYAML==5.4.1
pyzmq==22.2.1
readme-renderer==29.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rope==0.19.0
s3transfer==0.5.0
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
smmap==4.0.0
snowballstemmer==2.1.0
sqlparse==0.4.2
stevedore==3.4.0
stripe==2.64.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.3
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
twine==3.4.2
typing-extensions==3.10.0.2
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
zipp==3.5.0
